requirements.txt
alembic==1.8.0
bcrypt==3.2.2
blinker==1.4
cffi==1.15.0
click==8.1.3
dominate==2.6.0
Flask==2.1.0
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
Flask-WTF==1.0.1
greenlet==1.1.2
importlib-metadata==4.11.4
importlib-resources==5.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.0
MarkupSafe==2.1.1
Pillow==9.1.1
pycparser==2.21
SQLAlchemy==1.3.8
visitor==0.1.3
Werkzeug==2.0.0
WTForms==3.0.1
zipp==3.8.0
email_validator==1.2.1
python-dotenv==0.20.0